Title: 이차원 배열의 오름차순 정리, Time: 133ms, Memory: 8MB, Status: Passed - CodeTree

+++ 240603/이차원 배열의 오름차순 정리/ascending-order-of-two-dimensional-array.java
@@ -6,17 +6,17 @@ public class Main {
     public static void main(String[] args) throws Exception {
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
 
-        int n = Integer.parseInt(reader.readLine()); // 격자의 크기
-        int k = Integer.parseInt(reader.readLine()); // 숫자의 번호
+        long n = Long.parseLong(reader.readLine()); // 격자의 크기
+        long k = Long.parseLong(reader.readLine()); // 숫자의 번호
         
-        int left = 1; // 구간의 시작
-        int right = n * n; // 구간의 끝
+        long left = 1; // 구간의 시작
+        long right = n * n; // 구간의 끝
         
         while (left <= right) {
-            int mid = (left + right) / 2; // 중앙값
+            long mid = (left + right) / 2; // 중앙값
             
             // mid보다 작거나 같은 수의 개수 계산
-            int count = 0;
+            long count = 0;
             for (int idx = 1; idx <= n; idx++) {
                 count += Math.min(mid / idx, n);
             }
